+++ src/main/java/com/froggylord/constellation/config/VisualConfig.java
@@ -1,0 +1,7 @@
+package com.froggylord.constellation.config;
+
+public final class VisualConfig extends BaseConfigGroup {
+    public boolean titleBackdrop = true;
+    public boolean titleButtons = true;
+    public boolean reducedMotion = false;
+}